rename package gym to fingym
