+++ README.rst
@@ -43,7 +43,7 @@ You can perform a minimal install of ``fingym`` with:
     cd fingym
     pipenv shell
     pipenv install
-    export PYTHONPATH=$PYTHONPATH:/path/to/fingym/gym
+    export PYTHONPATH=$PYTHONPATH:/path/to/fingym/fingym
 
 Environments
 ============
@@ -76,7 +76,7 @@ Testing
 
 If you cloned this repo, add fingym to python path:
 
->> export PYTHONPATH=$PYTHONPATH:/path/to/fingym/gym
+>> export PYTHONPATH=$PYTHONPATH:/path/to/fingym/fingym
 
 We are using `pytest <http://doc.pytest.org>`_ for tests. You can run them via:
 
@@ -91,3 +91,4 @@ What's new
 ==========
 - 2020-02-05: First release. 3 year spy intraday minute steps. 10 year daily steps.
 - 2020-02-26: More environments from different symbols.
+- 2020-04-14: Renamed package from `gym` to `fingym`
