+++ fingym/data/filtered_aapl_data_10_yrs.csv
