+++ fingym/data/filtered_amd_data_10_yrs.csv
