+++ fingym/data/filtered_amzn_full_weekly.csv
