+++ fingym/data/filtered_cgc_data.csv
