+++ fingym/data/filtered_spy_data_10_yrs.csv
