+++ tests/amd_envs_test.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from gym.envs.amd_envs import AmdDailyEnv
+from fingym.envs.amd_envs import AmdDailyEnv
 
 import pytest
